Allow structs without public fields to implement Setter
If a struct does not provide a public field that might be set by
cleanenv, then the struct was ignored. This did make it impossible to
implement a custom struct with SetValue.

The change checks if a field implements SetValue. If SetValue is
implemented we stop recursing into the type to looks for other fields
that might be populated via cleanenv.